mcp-py/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.47.2",
    "groq>=0.24.0",
    "httpx>=0.28.1",
    "mcp>=1.7.1",
    "python-dotenv>=1.1.0",
]
